feat: active under bar 추가, active item 스크롤링 추가
